requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
lightning-utilities==0.11.8
loguru==0.7.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
torchmetrics==1.5.2
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
xxhash==3.5.0
yarl==1.17.1
